requirements.txt
requests
tenacity
python-dateutil
pyyaml 